# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tests Platform API",
    "description": "Assembly + administration engine for a large testing program.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/forms/compare": {
      "post": {
        "tags": [
          "forms"
        ],
        "summary": "Compare Forms",
        "description": "Cross-form comparability / equating-evidence report over a set of forms.\n\nShows whether the forms match by design on the canonical D=1 IRT scale (TIF / SE\n/ TCC overlay + per-\u03b8 dispersion + pass/flag). Comparability evidence only \u2014 not\nresponse-data equating (see the report's scope_note).",
        "operationId": "compare_forms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparabilityReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/{form_id}/lifecycle": {
      "get": {
        "tags": [
          "forms"
        ],
        "summary": "Get Form Lifecycle",
        "operationId": "get_form_lifecycle",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormLifecycle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/{form_id}/transition": {
      "post": {
        "tags": [
          "forms"
        ],
        "summary": "Transition Form",
        "description": "Move a form through the review/approve/publish lifecycle (records sign-off).",
        "operationId": "transition_form",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormLifecycle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/{form_id}/qa-report": {
      "get": {
        "tags": [
          "forms"
        ],
        "summary": "Get Form Qa Report",
        "description": "Server-side form-QA report (answer key, key balance, coverage, SE/TCC, TIF).",
        "operationId": "get_form_qa_report",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormQAReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/{form_id}/cross-validate": {
      "post": {
        "tags": [
          "forms"
        ],
        "summary": "Cross Validate Form",
        "description": "Validate an assembled form against the eatATA R oracle (read-only).\n\nRecompiles the form's blueprint+pool to the same canonical D=1 problem, solves\nit with eatATA via the oracle-r service, and compares to the stored OR-Tools\nresult. Never builds a deliverable form. Scope: single-form unweighted minimax\n(the eatATA bridge's objective).",
        "operationId": "cross_validate_form",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrossValidationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/{form_id}": {
      "get": {
        "tags": [
          "forms"
        ],
        "summary": "Get Form",
        "operationId": "get_form",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/{form_id}/tif-curve": {
      "get": {
        "tags": [
          "forms"
        ],
        "summary": "Get Form Tif Curve",
        "description": "Dense actual TIF over a theta grid (computed via the canonical metric).\n\nLets the UI draw a smooth actual curve against the discrete blueprint target.",
        "operationId": "get_form_tif_curve",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "theta_min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": -3.0,
              "title": "Theta Min"
            }
          },
          {
            "name": "theta_max",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 3.0,
              "title": "Theta Max"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 400,
              "minimum": 2,
              "default": 61,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TIFCurve"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/{form_id}/simulate": {
      "get": {
        "tags": [
          "forms"
        ],
        "summary": "Simulate Form",
        "description": "Simulate an examinee at ``theta`` walking this form (real engine + 2PL).",
        "operationId": "simulate_form",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "theta",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "examinee's true theta",
              "default": 0.0,
              "title": "Theta"
            },
            "description": "examinee's true theta"
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparabilityReport": {
        "properties": {
          "form_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Form Ids"
          },
          "n_forms": {
            "type": "integer",
            "title": "N Forms"
          },
          "metric": {
            "type": "string",
            "title": "Metric",
            "default": "logistic-D1-slope-intercept"
          },
          "theta_grid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Theta Grid"
          },
          "tolerance": {
            "type": "number",
            "title": "Tolerance"
          },
          "score_tolerance": {
            "type": "number",
            "title": "Score Tolerance"
          },
          "target": {
            "items": {
              "$ref": "#/components/schemas/TargetPoint"
            },
            "type": "array",
            "title": "Target"
          },
          "forms": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__comparability__FormSummary"
            },
            "type": "array",
            "title": "Forms"
          },
          "dispersion": {
            "items": {
              "$ref": "#/components/schemas/DispersionPoint"
            },
            "type": "array",
            "title": "Dispersion"
          },
          "max_tif_deviation": {
            "type": "number",
            "title": "Max Tif Deviation"
          },
          "max_tif_deviation_theta": {
            "type": "number",
            "title": "Max Tif Deviation Theta"
          },
          "max_expected_score_diff": {
            "type": "number",
            "title": "Max Expected Score Diff"
          },
          "max_expected_score_diff_theta": {
            "type": "number",
            "title": "Max Expected Score Diff Theta"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          },
          "scope_note": {
            "type": "string",
            "title": "Scope Note",
            "default": "Comparability evidence: forms matched by design on the IRT scale (canonical D=1). NOT statistical equating \u2014 no score-conversion tables are derived from examinee response data (that is downstream post-administration equating)."
          }
        },
        "type": "object",
        "required": [
          "form_ids",
          "n_forms",
          "theta_grid",
          "tolerance",
          "score_tolerance",
          "target",
          "forms",
          "dispersion",
          "max_tif_deviation",
          "max_tif_deviation_theta",
          "max_expected_score_diff",
          "max_expected_score_diff_theta",
          "passed",
          "flags"
        ],
        "title": "ComparabilityReport"
      },
      "ComparabilityRequest": {
        "properties": {
          "form_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Form Ids"
          },
          "tolerance": {
            "type": "number",
            "title": "Tolerance",
            "default": 1.0
          },
          "score_tolerance": {
            "type": "number",
            "title": "Score Tolerance",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "form_ids"
        ],
        "title": "ComparabilityRequest"
      },
...
      "CrossValidationResult": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "package": {
            "type": "string",
            "title": "Package"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "ortools": {
            "$ref": "#/components/schemas/CrossValSide"
          },
          "oracle": {
            "$ref": "#/components/schemas/CrossValOracle"
          },
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CrossValComparison"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "package",
          "ortools",
          "oracle"
        ],
        "title": "CrossValidationResult",
        "description": "Full cross-validation payload for the UI.\n\n``status``: ``ok`` (ran and compared), ``unsupported`` (blueprint outside the\noracle's scope \u2014 single-form unweighted minimax only), ``oracle_unavailable``\n(service unreachable), or ``error``."
      },
...
      "FormLifecycle": {
        "properties": {
          "form_id": {
            "type": "string",
            "title": "Form Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "frozen": {
            "type": "boolean",
            "title": "Frozen"
          },
          "available_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Actions"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ReviewEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "form_id",
          "state",
          "frozen",
          "available_actions",
          "events"
        ],
        "title": "FormLifecycle"
      },
      "FormQAReport": {
        "properties": {
          "form_id": {
            "type": "string",
            "title": "Form Id"
          },
          "lifecycle_state": {
            "type": "string",
            "title": "Lifecycle State"
          },
          "n_items": {
            "type": "integer",
            "title": "N Items"
          },
          "metric": {
            "type": "string",
            "title": "Metric",
            "default": "logistic-D1-slope-intercept"
          },
          "answer_key": {
            "items": {
              "$ref": "#/components/schemas/AnswerKeyEntry"
            },
            "type": "array",
            "title": "Answer Key"
          },
          "key_balance": {
            "$ref": "#/components/schemas/KeyBalance"
          },
          "coverage": {
            "items": {
              "$ref": "#/components/schemas/CoverageRow"
            },
            "type": "array",
            "title": "Coverage"
          },
          "curve": {
            "items": {
              "$ref": "#/components/schemas/QAPsychometricPoint"
            },
            "type": "array",
            "title": "Curve"
          },
          "marginal_reliability": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Marginal Reliability"
          },
          "tif_actual_vs_target": {
            "items": {
              "$ref": "#/components/schemas/TIFActualTarget"
            },
            "type": "array",
            "title": "Tif Actual Vs Target"
          }
        },
        "type": "object",
        "required": [
          "form_id",
          "lifecycle_state",
          "n_items",
          "answer_key",
          "key_balance",
          "coverage",
          "curve",
          "tif_actual_vs_target"
        ],
        "title": "FormQAReport"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TargetPoint": {
        "properties": {
          "theta": {
            "type": "number",
            "title": "Theta"
          },
          "target": {
            "type": "number",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "theta",
          "target"
        ],
        "title": "TargetPoint"
      },
...
      "TransitionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor": {
            "type": "string",
            "title": "Actor",
            "default": "anonymous"
          },
          "actor_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Role"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "TransitionRequest",
        "description": "A requested lifecycle transition. ``actor``/``actor_role`` are the *claimed*\nidentity (recorded; not yet authorization-checked \u2014 role hook is a stub)."
      },
...
    }
  }
}